pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-book-digitizer"
version = "0.1.0"
description = "Digitize scanned PDF books into OCR text, extracted images, and Calibre-friendly HTML."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ollama>=0.4.7",
  "Pillow>=10.3.0",
  "PyMuPDF>=1.24.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
digitize-book = "pdf_book_digitizer.cli:main"
digitize-images = "pdf_book_digitizer.cli_images:main"
extract-pdf-pages = "pdf_book_digitizer.cli_extract_pages:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
